Moved Interpreter Code into it's own class
